move ai to shared dir
