+++ packages/vscode-extension/src/ai/mcp/configCreator.ts
