+++ packages/vscode-extension/src/ai/mcp/index.test.ts
