+++ packages/vscode-extension/src/ai/mcp/models.ts
