+++ packages/vscode-extension/src/ai/mcp/utils.ts
